Develop

+++ Tests/TestQuagga.py
@@ -1,14 +1,14 @@
 from unittest import TestCase
 
 from Quagga import Quagga, EmailDirectoryReader, ModelBuilder
-from Quagga.Utils.Email import EmailMessage
 import json
 import os
 import filecmp
-from email import parser as ep
 
+from Tests.TestUtils import eq
 
 class TestQuagga(TestCase):
+	_multiprocess_shared_ = True
 	def get_relative_filename(self, file):
 		dirname = os.path.dirname(__file__)
 		filename = os.path.join(dirname, file)
@@ -58,8 +58,9 @@ PL
 		self.test_expected_dir = self.test_data_dir + '/output_expected'
 		self.quagga = Quagga(EmailDirectoryReader(self.test_data_dir), self.test_output_dir)
 
-		for f in os.listdir(self.test_output_dir):
-			os.remove(os.path.join(self.test_output_dir, f))
+		if os.path.isdir(self.test_output_dir):
+			for f in os.listdir(self.test_output_dir):
+				os.remove(os.path.join(self.test_output_dir, f))
 
 		self.test_mails = ["""That's it.  Thanks to plove I am no longer entering my own deals.
 
@@ -139,39 +140,26 @@ x3-0977 """]
 	def output_filename(self, email_input):
 		return self.test_output_dir + '/' + email_input.filename
 
-	@staticmethod
-	def assertComparisionError(expected, actual):
-		TestQuagga.printComparisionError(expected, actual)
-		assert expected == actual
-
-	@staticmethod
-	def printComparisionError(expected, actual):
-		if expected != actual:
-			print("expected output")
-			print(expected)
-			print("actual parsed output")
-			print(actual)
-
 	def test_emails_input(self):
 		for email_input in self.quagga.emails_input:
 			with open(self.expected_filename(email_input) + self.quagga.fileending_input, 'r') as fp:
 				expected_input = json.loads(fp.read())[Quagga.INPUT_NAME]
-				TestQuagga.assertComparisionError(email_input.file, expected_input['file'])
-				TestQuagga.assertComparisionError(email_input.folder, expected_input['folder'])
-				TestQuagga.assertComparisionError(email_input.sent_string, expected_input['sent'])
-				TestQuagga.assertComparisionError(email_input.id, expected_input['id'])
-				TestQuagga.assertComparisionError(email_input.mailbox, expected_input['mailbox'])
-				TestQuagga.assertComparisionError(email_input.subject, expected_input['subject'])
-				TestQuagga.assertComparisionError(email_input.sender, expected_input['sender'])
-				TestQuagga.assertComparisionError(email_input.xsender, expected_input['xsender'])
-				TestQuagga.assertComparisionError(email_input.to, expected_input['to'])
-				TestQuagga.assertComparisionError(email_input.xto, expected_input['xto'])
-				TestQuagga.assertComparisionError(email_input.cc, expected_input['cc'])
-				TestQuagga.assertComparisionError(email_input.xcc, expected_input['xcc'])
-				TestQuagga.assertComparisionError(email_input.bcc, expected_input['bcc'])
-				TestQuagga.assertComparisionError(email_input.xbcc, expected_input['xbcc'])
-				TestQuagga.assertComparisionError(email_input.body, expected_input['body'])
-				TestQuagga.assertComparisionError(email_input.clean_body, expected_input['clean_body'])
+				eq(email_input.file, expected_input['file'])
+				eq(email_input.folder, expected_input['folder'])
+				eq(email_input.sent_string, expected_input['sent'])
+				eq(email_input.id, expected_input['id'])
+				eq(email_input.mailbox, expected_input['mailbox'])
+				eq(email_input.subject, expected_input['subject'])
+				eq(email_input.sender, expected_input['sender'])
+				eq(email_input.xsender, expected_input['xsender'])
+				eq(email_input.to, expected_input['to'])
+				eq(email_input.xto, expected_input['xto'])
+				eq(email_input.cc, expected_input['cc'])
+				eq(email_input.xcc, expected_input['xcc'])
+				eq(email_input.bcc, expected_input['bcc'])
+				eq(email_input.xbcc, expected_input['xbcc'])
+				eq(email_input.body, expected_input['body'])
+				eq(email_input.clean_body, expected_input['clean_body'])
 
 	def test_store_input(self):
 		self.quagga.store_input(self.test_output_dir)
@@ -182,18 +170,17 @@ x3-0977 """]
 	def test__build_model(self):
 		self.quagga.model = None
 		self.quagga._build_model(ModelBuilder(with_crf=False, zones=5, trainset='asf'))
-		assert self.quagga.model.with_crf == False
-		assert self.quagga.model.zones == 5
-		assert self.quagga.model.trainset == 'asf'
+		eq(self.quagga.model.with_crf, False)
+		eq(self.quagga.model.zones, 5)
+		eq(self.quagga.model.trainset, 'asf')
 
 	def test__predict(self):
-		assert self.quagga._predict(self.test_mails[0]) == self.test_predict_expected[0]
+		eq(self.quagga._predict(self.test_mails[0]), self.test_predict_expected[0])
 
 	def test_emails_predicted(self):
 		for email_predicted, email_expected in zip(self.quagga.emails_predicted(input_reader=self.test_mails),
 		                                           self.test_predict_expected):
-			TestQuagga.printComparisionError(email_expected, email_predicted)
-			assert email_predicted == email_expected
+			eq(email_predicted, email_expected)
 
 	def test_store_predicted(self):
 		self.quagga.store_predicted(self.test_output_dir)
